+++ spec/brainstem/presenter_spec.rb
@@ -161,19 +161,35 @@ describe Brainstem::Presenter do
         end
 
         @presenter = some_presenter.new
-        @post = Post.first
       end
 
-      it "outputs the associated object's id and type" do
-        data = @presenter.present_and_post_process(@post)
-        expect(data[:subject_id]).to eq(@post.subject_id.to_s)
-        expect(data[:subject_type]).to eq(@post.subject_type)
+      let(:presented_data) { @presenter.present_and_post_process(post) }
+
+      context "when polymorphic association exists" do
+        let(:post) { Post.find(1) }
+
+
+        it "outputs the object as a hash with the id & class table name" do
+          expect(presented_data[:subject_ref]).to eq({ :id => post.subject.id.to_s,
+                                                       :key => post.subject.class.table_name })
+        end
+
+        it "outputs custom names for the object as a hash with the id & class table name" do
+          expect(presented_data[:another_subject_ref]).to eq({ :id => post.subject.id.to_s,
+                                                               :key => post.subject.class.table_name })
+        end
       end
 
-      it "outputs custom names for an associated object's id and type" do
-        data = @presenter.present_and_post_process(@post)
-        expect(data[:another_subject_id]).to eq(@post.subject_id.to_s)
-        expect(data[:another_subject_type]).to eq(@post.subject_type)
+      context "when polymorphic association does not exist" do
+        let(:post) { Post.find(3) }
+
+        it "outputs nil" do
+          expect(presented_data[:subject_ref]).to be_nil
+        end
+
+        it "outputs nil" do
+          expect(presented_data[:another_subject_ref]).to be_nil
+        end
       end
     end